Mac has something like vmstat
I don't use this myself so I'm not sure, but it appears that Mac has a `vm_stat` which is analogous to `vmstat`, though the outputs seem quite differently formatted.  (FreeBSD also seems to have `vmstat` so I'm not sure where in the BSD genealogy each of these were introduced.)  Hope you find this a helpful suggestion to add.